dump/[202412] zuto/pyproject.toml
[project]
name = "zuto"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ZackaryW", email = "36378555+ZackaryW@users.noreply.github.com" }
]
dependencies = [
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "zuu>=3.0.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/zuto"]
